Add example of using <details> tag in Markdown kitchen sink (#1297)
* Add example of using <details> tag in Markdown kitchen sink

I hoped you would consider adding this example of how to use the <details> tag in the Markdown kitchen sink. I found some useful information in https://github.com/just-the-docs/just-the-docs/issues/246 and, this an example would be handy on this page.

* removed style and called it "collapsed section" instead of dropdown

Also added a link to GitHub's description of how to write a collapsed section.